리뷰 수정 삭제 기능 추가구현

+++ back-end/sql/review/delete.js
@@ -1,18 +1,20 @@
 const mysql = require("../../connection/mysqlConnection");
 
-function deleteReview(reviewId) {
+function deleteReview(reviewId, userId) {
   return new Promise((resolve, reject) => {
-    const query = "DELETE FROM review WHERE review_id = ?";
-    mysql.query(query, [reviewId], (err, result) => {
+    const query = "DELETE FROM review WHERE review_id = ? AND user_id = ?";
+    mysql.query(query, [reviewId, userId], (err, result) => {
       if (err) {
         reject(err);
       } else {
-        resolve(result);
+        if (result.affectedRows > 0) {
+          resolve({ message: "SUCCESS" });
+        } else {
+          resolve({ message: "NO_REVIEW_DELETED" });
+        }
       }
     });
   });
 }
 
-module.exports = {
-  deleteReview,
-};
+module.exports = { deleteReview };
